+++ Tetris/GlobalValues.hpp
@@ -1,9 +1,9 @@
 #pragma once
 
-constexpr int NUMBER_OF_ROWS = 20;
-constexpr int NUMBER_OF_COLUMNS = 10;
-constexpr int GRID = 25;
-constexpr int GAME_SPEED = 200;
+constexpr uint8_t NUMBER_OF_ROWS = 20;
+constexpr uint8_t NUMBER_OF_COLUMNS = 10;
+constexpr uint8_t GRID = 25;
+constexpr uint8_t GAME_SPEED = 200;
 
 enum class FieldState
 {
