+++ Tetris/blockBoard.cpp
@@ -7,7 +7,7 @@ BlockBoard::BlockBoard() noexcept
     {
         for (int j = 0; j < NUMBER_OF_ROWS; j++)
         {
-            blockBoard_.at(i).at(j).block_.setPosition(i * GRID + GRID, j * GRID + GRID);
+            blockBoard_.at(i).at(j).block_.setPosition(static_cast<float>(i * GRID + GRID), static_cast<float>(j * GRID + GRID));
             blockBoard_.at(i).at(j).block_.setSize(Vector2f(GRID, GRID));
             blockBoard_.at(i).at(j).block_.setFillColor(Color::White);
         }
